Move back from snaps-utils
