+++ packages/snaps-utils/src/index.ts
@@ -20,4 +20,3 @@ export * from './snaps';
 export * from './types';
 export * from './versions';
 export * from './virtual-file';
-export * from './registry';
